Mapping docker container's out directory with host's
The build script creates an out directory under the working directory
(which can be mounted from the host) and moves the compiled kernel
artifacts there. When the container is run with a volume mount
(for example, using -v $(pwd)/out:/build/out), the host’s out
directory will contain the final packaged kernel as well as any
intermediate build outputs.